Rename ops to logical_ops
