+++ src/daft-plan/src/lib.rs
@@ -3,8 +3,8 @@
 mod builder;
 mod display;
 mod join;
+mod logical_ops;
 mod logical_plan;
-mod ops;
 mod optimization;
 mod partitioning;
 mod physical_ops;
