+++ src/daft-plan/src/logical_ops/coalesce.rs
